app/requirements.txt
openpyxl>=3.0.9
pycurl>=7.43.0.6
validators>=0.18.2
python-magic>=0.4.24
pyinstaller>=4.7
Flask>=Flask-2.0.2
